Remove CleanupHackLevel from CGException.
The CleanupHackLevel was an hack to avoid misoptimization
to the old exception handling intrinsics, which have several
issues with function inliner.

The landingpad and resume instructions are added to LLVM IR
since LLVM 3.0.  Now, we have a canonical way to generate
exception handling code.

Besides, the `CleanupHackLevel` is a compile-time constant,
thus the other alternatives has been considered as dead code
for a while.

This patch removes the dead code, and refines the
getEHResumeBlock() slightly.